pyproject.toml
[project]
name = "kakao-maps"
version = "0.0.4"
description = "MCP for Kakao Maps"
authors = [{name = "clauzewitz"}]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.10.6",
    "aiohttp>=3.12.14",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kakao_maps"]

[project.scripts]
genesis-connect = "kakao_maps.server:main"